Convert isize to i32

This changes pretty much all of tcod's API to use i32 instead of int/isize, so it's a pretty big breaking change.

+++ examples/astar-path-finding.rs
@@ -1,11 +1,9 @@
-#![feature(unboxed_closures)]
-
 extern crate tcod;
 
 use tcod::AStarPath;
 
 fn create_path() -> AStarPath<'static> {
-    let chess_board: [[isize; 8]; 8] = [
+    let chess_board: [[i32; 8]; 8] = [
         [1, 0, 1, 0, 1, 0, 1, 0],
         [0, 1, 0, 1, 0, 1, 0, 1],
         [1, 0, 1, 0, 1, 0, 1, 0],
@@ -16,7 +14,7 @@ fn create_path() -> AStarPath<'static> {
         [0, 1, 0, 1, 0, 1, 0, 1],
     ];
     // Movement like in Checkers: you can only move to the square of the same colour
-    let can_move = move |&mut: from: (isize, isize), to: (isize, isize)| -> f32 {
+    let can_move = move |&mut: from: (i32, i32), to: (i32, i32)| -> f32 {
         let (fx, fy) = from;
         let (tx, ty) = to;
         if chess_board[fy as usize][fx as usize] == chess_board[ty as usize][tx as usize] {
